ScriptingGlue/Python: Make unicode string creation more robust.
The old implementation expected UTF-8 encoded strings, and
could return NULL on invalid UTF-8, e.g., when the string
was in a different encoding.

Now we check for such failures and retry converting the
string assuming it is encoded in the current system locale,
which might be valid for some strings returned by PsychHID
on Windows. On Python 2 we fallback to old style strings,
as it lacks the locale sensitive decoding.

As Python seems to have no sense of backwards compatibility,
with incompatible changes from Python 3.0 -> 3.2, then again
for 3.2 -> 3.6 and again from 3.6 -> 3.7, sometimes even only
affecting one OS, e.g., Windows, we implement any half-way
meaningful fallback of trying to decode.

If everything else fails, we assume Latin-1 encoding to either
get a valid Latin-1 string, or something that is essentially
the raw Byte string with values 0 - 255 which may be suitable
for debugging or further workarounds from the calling Python
script.

In any case, don't return NULL, as that leads to crashes,
but at least return some "NULL" or "Invalid" string.

-> This is meant to work around issues found when calling
   PsychHID('Devices', 4) on a japanese MS-Windows 10, which
   may encode device product name strings in locale encoding
   of codepage 932 (similar to Shift_JIS?). My suspicion now,
   after looking at CPython master's unicode implementation
   for the better part of a day, is that that system may just
   return a truncated, incomplete, and thereby broken multi-byte
   product string, as it seems to trip up the locale decoder
   as well and "incomplete multibyte encoded wchar sequence"
   seems to be the only way to make that function fail. Cfe.
   Microsoft docs for mbrtowc() - return code -2, which is
   treated as complete failure by Python 3.8 devel master,
   whereas the MS docs suggest it just means the last character
   might be incompletely specified. That said, the error
   was reported against Python 3.6.1, and Python 3.7 has
   different behaviour. Testing of this patch is pending,
   which has more fallbacks than a previous iteration.

   Reference: https://github.com/psychopy/psychopy/issues/2453